+++ cobaya/samplers/evaluate/evaluate.py
@@ -12,7 +12,7 @@ from typing import Mapping
 
 # Local
 from cobaya.sampler import Sampler
-from cobaya.collection import Collection
+from cobaya.collection import SampleCollection
 from cobaya.log import LoggedError
 
 
@@ -30,8 +30,7 @@ class evaluate(Sampler):
             raise LoggedError(
                 self.log,
                 "Could not convert the number of samples to an integer: %r", self.N)
-        self.one_point = Collection(
-            self.model, self.output, name="1")
+        self.one_point = SampleCollection(self.model, self.output, name="1")
         self.log.info("Initialized!")
 
     def run(self):
@@ -48,9 +47,9 @@ class evaluate(Sampler):
                 self.log.info("Evaluating sample #%d ------------------------------",
                               i + 1)
             self.log.info("Looking for a reference point with non-zero prior.")
-            reference_point = self.model.prior.reference(random_state=self._rng)
+            reference_values = self.model.prior.reference(random_state=self._rng)
             reference_point = dict(
-                zip(self.model.parameterization.sampled_params(), reference_point))
+                zip(self.model.parameterization.sampled_params(), reference_values))
             for p, v in (self.override or {}).items():
                 if p not in reference_point:
                     raise LoggedError(
@@ -91,6 +90,6 @@ class evaluate(Sampler):
         Auxiliary function to define what should be returned in a scripted call.
 
         Returns:
-           The sample ``Collection`` containing the sequentially discarded live points.
+           The sample ``SampleCollection`` containing the sequentially discarded live points.
         """
         return {"sample": self.one_point}